% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funstattest_authors.R
\name{funStatTest_authors}
\alias{funStatTest_authors}
\title{List \code{funStatTest} package contributors}
\usage{
funStatTest_authors()
}
\description{
List \code{funStatTest} package contributors
}
\author{
Zaineb Smida, Ghislain DURIF, Lionel Cucala
}
\keyword{internal}
